Create CodeForces_Masha and a Beautiful Tree.cpp
Problem Title: 1741D  Masha and a Beautiful Tree
The girl named Masha was walking in the forest and found a complete binary tree of height n and a permutation p of length m=2n.
A complete binary tree of height n is a rooted tree such that every vertex except the leaves has exactly two sons, and the length of the path from the root to any of the leaves is n. The picture below shows the complete binary tree for n=2.
A permutation is an array consisting of n different integers from 1 to n. For example, [2,3,1,5,4] is a permutation, but [1,2,2] is not (2 occurs twice), and [1,3,4] is also not a permutation (n=3, but there is 4 in the array).